+++ app/src/main/java/com/lacrose/lc/lclacrose/Model/PavimentoLotes.java
@@ -21,7 +21,7 @@ public  class PavimentoLotes {
     public Long data = null;
     public Long dataCreate = null;
     public Long dataFab = null;
-    public HashMap<String, Integer> dimenssion;
+    public HashMap<String, Double> dimenssion;
     public String fabricante = "";
     public Double fpk = null;
     public Long hora = null;
@@ -117,11 +117,11 @@ public  class PavimentoLotes {
         this.dataFab = dataFab;
     }
 
-    public HashMap<String, Integer> getDimenssion() {
+    public HashMap<String, Double> getDimenssion() {
         return dimenssion;
     }
 
-    public void setDimenssion(HashMap<String, Integer> dimenssion) {
+    public void setDimenssion(HashMap<String, Double> dimenssion) {
         this.dimenssion = dimenssion;
     }
