+++ app/src/main/java/com/lacrose/lc/lclacrose/Model/PrismaLotes.java
@@ -23,7 +23,7 @@ public  class PrismaLotes {
     public Long dataCreate = null;
     public Long dataFab = null;
     public Long dataGraute = null;
-    public HashMap<String, Integer> dimenssion;
+    public HashMap<String, Double> dimenssion;
     public String fabricante = "";
     public String funcEstrutural = "";
     public Double fpk = null;
@@ -137,11 +137,11 @@ public  class PrismaLotes {
         this.dataGraute = dataGraute;
     }
 
-    public HashMap<String, Integer> getDimenssion() {
+    public HashMap<String, Double> getDimenssion() {
         return dimenssion;
     }
 
-    public void setDimenssion(HashMap<String, Integer> dimenssion) {
+    public void setDimenssion(HashMap<String, Double> dimenssion) {
         this.dimenssion = dimenssion;
     }
