+++ app/src/main/java/com/lacrose/lc/lclacrose/PavimentoMoldActivity.java
@@ -302,27 +302,27 @@ public class PavimentoMoldActivity extends MainActivity implements DatePickerDia
         if(validateFields()){
             final CollectionReference ref_lote = database.collection(getString(R.string.lote_tag));
             newLote = new PavimentoLotes();
-            HashMap<String, Integer> dimenssionHash = new HashMap<>();
+            HashMap<String, Double> dimenssionHash = new HashMap<>();
             if(!String.valueOf(spinner_dimenssion.getSelectedItem()).equals(getString(R.string.dimenssion_prompt))) {
                 if (String.valueOf(spinner_dimenssion.getSelectedItem()).equals(getString(R.string.d60_100_200))) {
-                    dimenssionHash.put(getString(R.string.largura), 60);
-                    dimenssionHash.put(getString(R.string.altura), 100);
-                    dimenssionHash.put(getString(R.string.comprimento), 200);
+                    dimenssionHash.put(getString(R.string.largura), 60.0);
+                    dimenssionHash.put(getString(R.string.altura), 100.0);
+                    dimenssionHash.put(getString(R.string.comprimento), 200.0);
                     newLote.setDimenssion(dimenssionHash);
                 } else if (String.valueOf(spinner_dimenssion.getSelectedItem()).equals(getString(R.string.d80_100_200))) {
-                    dimenssionHash.put(getString(R.string.largura), 80);
-                    dimenssionHash.put(getString(R.string.altura), 100);
-                    dimenssionHash.put(getString(R.string.comprimento), 200);
+                    dimenssionHash.put(getString(R.string.largura), 80.0);
+                    dimenssionHash.put(getString(R.string.altura), 100.0);
+                    dimenssionHash.put(getString(R.string.comprimento), 200.0);
                     newLote.setDimenssion(dimenssionHash);
                 } else if (String.valueOf(spinner_dimenssion.getSelectedItem()).equals(getString(R.string.d100_100_200))) {
-                    dimenssionHash.put(getString(R.string.largura), 100);
-                    dimenssionHash.put(getString(R.string.altura), 100);
-                    dimenssionHash.put(getString(R.string.comprimento), 200);
+                    dimenssionHash.put(getString(R.string.largura), 100.0);
+                    dimenssionHash.put(getString(R.string.altura), 100.0);
+                    dimenssionHash.put(getString(R.string.comprimento), 200.0);
                     newLote.setDimenssion(dimenssionHash);
                 } else if(String.valueOf(spinner_dimenssion.getSelectedItem()).equals(getString(R.string.outras_dim))){
-                    dimenssionHash.put(getString(R.string.largura), Integer.parseInt(edit_largra.getText().toString()));
-                    dimenssionHash.put(getString(R.string.altura), Integer.parseInt(edit_altura.getText().toString()));
-                    dimenssionHash.put(getString(R.string.comprimento), Integer.parseInt(edit_comprimeto.getText().toString()));
+                    dimenssionHash.put(getString(R.string.largura), Double.parseDouble(edit_largra.getText().toString()));
+                    dimenssionHash.put(getString(R.string.altura), Double.parseDouble(edit_altura.getText().toString()));
+                    dimenssionHash.put(getString(R.string.comprimento), Double.parseDouble(edit_comprimeto.getText().toString()));
                     newLote.setDimenssion(dimenssionHash);
                 }
             }
